server/requirements.txt
Flask==2.3.2
pandas==2.0.1
python-dotenv==1.0.0
